+++ DMeServices.DAL/BldPermitsAttachments.cs
@@ -20,13 +20,13 @@ namespace DMeServices.DAL
             this.BldPermitsAttachmentDetails = new HashSet<BldPermitsAttachmentDetails>();
         }
     
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string AttachmentName { get; set; }
         public string AttachmentContentType { get; set; }
         public System.DateTime InsertDate { get; set; }
         public string AttachmentPath { get; set; }
         public string Description { get; set; }
-        public long BldPermitId { get; set; }
+        public int BldPermitId { get; set; }
         public byte[] AttachmentStream { get; set; }
         public string CreatedBy { get; set; }
         public Nullable<System.DateTime> CreatedOn { get; set; }
