+++ DMeServices.Models/Models/BuildingServices/BuildingSupervision.cs
@@ -11,10 +11,10 @@ namespace DMeServices.Models.BuildingServices
     public  class BuildingSupervision
     {
         public int ID { get; set; }
-        public long BldPermitID { get; set; }
+        public int BldPermitID { get; set; }
         public int ServiceTypeID { get; set; }
         public int Status { get; set; }
-        public long OwnerCivilId { get; set; }
+        public int OwnerCivilId { get; set; }
         public string KrokiNO { get; set; }
         public string OwnerName { get; set; }
         public string OwnerPhoneNo { get; set; }
@@ -25,7 +25,7 @@ namespace DMeServices.Models.BuildingServices
         public string LicenseNo { get; set; }
         public Nullable<int> DmEngineerNo { get; set; }
         public string DmEngineerComments { get; set; }
-        public Nullable<long> ConsultantCivilId { get; set; }
+        public Nullable<int> ConsultantCivilId { get; set; }
         public System.DateTime RequestDate { get; set; }
         public string TransactNo { get; set; }
         public string DmFileNumber { get; set; }
