+++ DMeServices.Models/Models/BuildingServices/Payments.cs
@@ -9,7 +9,7 @@ namespace DMeServices.Models.BuildingServices
     public class Payments
     {
         public int PaymentID { get; set; }
-        public Nullable<long> BldPermitId { get; set; }
+        public Nullable<int> BldPermitId { get; set; }
         public Nullable<decimal> PaymentTotalAmount { get; set; }
         public string TokenID { get; set; }
         public Nullable<System.DateTime> PaymentDate { get; set; }
