+++ DMeServices.Models/Models/BuildingServices/PermitsAttachmentDetails.cs
@@ -9,13 +9,13 @@ namespace DMeServices.Models.BuildingServices
    public class PermitsAttachmentDetails
     {
 
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string AttachmentName { get; set; }
         public string AttachmentContentType { get; set; }
         public System.DateTime InsertDate { get; set; }
         public string AttachmentPath { get; set; }
         public string Description { get; set; }
-        public long BldPermitAttId { get; set; }
+        public int BldPermitAttId { get; set; }
         public byte[] AttachmentStream { get; set; }
         public Nullable<int> DmEngineerNo { get; set; }
         public string CreatedBy { get; set; }
